fixes and additions

+++ DBAPI.Library.Example/Program.cs
@@ -9,7 +9,6 @@
         public static void Main(string[] args)
         {
             var converter = ObjectConverter.GetTypeConverter<MyDataType>();
-            Console.WriteLine(converter.MySqlTypesString);
             var apiclient = new APIClient(new APIClientParameters()
             {
                 Host = "localhost",
@@ -39,24 +38,37 @@
             {
                 ID = 1,
                 MyString = "real2",
+                MyEnum = MyEnum.Two,
+                MyDateTime = DateTime.UtcNow,
             };
-            try
+            var dateTime = DateTime.Now;
+            var myObject2 = new MyDataType()
             {
-                apiclient.Insert("testcol", myObject);
-            }
-            catch (ArgumentException e)
-            {
-                if (!e.Message.Contains("duplicate"))
-                    throw;
-            }
-            //var entity = apiclient.GetById<MyDataType>("testcol", 1);
-            var entity = apiclient.GetByFields<MyDataType>("testcol", ("MyString", "real2"), ("ID", 1)).FirstOrDefault();
+                ID = 2,
+                MyString = "obj3",
+                MyEnum = MyEnum.One,
+                MyDateTime = dateTime,
+            };
+            var collection = apiclient.GetCollection<MyDataType>("testcol");
+            if (collection.Any(x => x.ID == myObject.ID))
+                apiclient.DeleteById("testcol", myObject.ID);
+            if (collection.Any(x => x.ID == myObject2.ID))
+                apiclient.DeleteById("testcol", myObject2.ID);
+            apiclient.Insert("testcol", myObject);
+            apiclient.Insert("testcol", myObject2);
+            var entity = apiclient.GetById<MyDataType>("testcol", 1);
+            var entity2 = apiclient.GetByFields<MyDataType>("testcol", ("MyDateTime", $"{dateTime:yyyy-MM-dd HH:mm:ss}"), ("ID", 2))
+                ?.FirstOrDefault();
+            if (entity2 == null)
+                throw new Exception("Entity2 is null");
             Console.WriteLine(entity.MyString);
+            Console.WriteLine(entity2.MyEnum.ToString());
             entity.MyString = "real3";
             apiclient.Update("testcol", entity);
             entity = apiclient.GetById<MyDataType>("testcol", 1);
             Console.WriteLine(entity.MyString);
             apiclient.Delete("testcol", myObject);
+            apiclient.Delete("testcol", myObject2);
             apiclient.DeleteCollection("testcol");
 
             apiclient.Ping("test");
